misc: move around and rename flag file
