+++ lib/run.js
@@ -1,6 +1,6 @@
 const webpackCli = require("./webpack-cli");
 const webpackInstance = require("./instance");
-const { core, commands } = require("./descriptions/args-detailed");
+const { core, commands } = require("./utils/cli-flags");
 const cmdArgs = require("command-line-args");
 
 process.on("uncaughtException", error => {
